Note for the page
This is the central storage for my web project/development.
This is for self or personal use only not intended to share to any one or to sell.
However visitors may find some works similar to other developer, this because I have them as my resources and or
they might find some of my original works helpful so they just pattern their work with mine. 
Resources or sources will be posted or will be shown to Thanks To You document. 

by: Kojiro Musashi (Dec.8, 2020)